fix seeder create

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -15,7 +15,7 @@ class CreateUsersTable extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->integer('level_id')->default(1);
+            $table->unsignedBigInteger('level_id')->default(1);
             $table->enum('role',['admin','user'])->default('user');
             $table->string('full_name');
             $table->string('user_name')->unique();
